fhir-server-compare/requirements.txt
httpx>=0.27
PyYAML>=6.0
# Optional, only needed when GCP_FHIR_STORE_URL is set:
#   pip install google-auth
